Refactored stored procedures and triggers 	- Changed comments in stored procedures to make the wording 	  consistent with all other stored procedure files. 	- Fixed bugs in stored procedures where a check to see if the 	  user is authorized isn't made before updating or deleting an entity. 	- Fixed bug in trigger to delete a bank account. Now, instead of 	  an error being thrown if a bank account that's going to be deleted as transfers with another bank account, the bank account is set as inactive. 	- Changed an error message in a the trigger to update a bank sub 	  account to for better understanding of the error.

+++ sql/stored-procedures/bank-account/delete.sql
@@ -4,7 +4,7 @@ DROP PROCEDURE IF EXISTS `delete_user_bank_account`$$
 
 CREATE PROCEDURE `delete_user_bank_account` (IN param_userId INT UNSIGNED, param_bankAccountId INT UNSIGNED)
 BEGIN
-	SET @errorMessage = "Cannot delete a bank account that the user doesn't own";
+	SET @errorMessage = "Cannot delete a bank account that the user doesn't have";
 	SET @bankId = (SELECT bank_id FROM BankAccount WHERE id = param_bankAccountId);
 
 	-- If the bank account doesn't exist
@@ -19,6 +19,18 @@ BEGIN
 		SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = @errorMessage;
 	END IF;
 
-	DELETE FROM BankAccount WHERE id = param_bankAccountId;
+	SET @numOfTransfers = (
+		SELECT COUNT(id) FROM BankAccountTransfer
+			WHERE BankAccountTransfer.from_bank_account_id = param_bankAccountId OR 
+					BankAccountTransfer.to_bank_account_id = param_bankAccountId
+	);
+
+	-- If the bank account has transfers between other accounts (cannot be deleted but will be inactive)
+	IF @numOfTransfers > 0 THEN
+		UPDATE BankAccount SET active = FALSE WHERE id = param_bankAccountId;
+	ELSE
+		DELETE FROM BankAccount WHERE id = param_bankAccountId;
+	END IF;
+
 END$$
 DELIMITER ;
